[Refactor] refactor native analyse to support external analyze
