+++ fe/fe-core/src/main/java/com/starrocks/journal/JournalEntity.java
@@ -130,9 +130,9 @@ import com.starrocks.scheduler.persist.TaskRunStatusChange;
 import com.starrocks.sql.ast.UserIdentity;
 import com.starrocks.staros.StarMgrJournal;
 import com.starrocks.statistic.AnalyzeJob;
-import com.starrocks.statistic.AnalyzeStatus;
 import com.starrocks.statistic.BasicStatsMeta;
 import com.starrocks.statistic.HistogramStatsMeta;
+import com.starrocks.statistic.NativeAnalyzeStatus;
 import com.starrocks.system.Backend;
 import com.starrocks.system.ComputeNode;
 import com.starrocks.system.Frontend;
@@ -686,12 +686,12 @@ public class JournalEntity implements Writable {
                 break;
             }
             case OperationType.OP_ADD_ANALYZE_STATUS: {
-                data = AnalyzeStatus.read(in);
+                data = NativeAnalyzeStatus.read(in);
                 isRead = true;
                 break;
             }
             case OperationType.OP_REMOVE_ANALYZE_STATUS: {
-                data = AnalyzeStatus.read(in);
+                data = NativeAnalyzeStatus.read(in);
                 isRead = true;
                 break;
             }
