+++ fe/fe-core/src/main/java/com/starrocks/persist/EditLog.java
@@ -99,6 +99,7 @@ import com.starrocks.statistic.AnalyzeJob;
 import com.starrocks.statistic.AnalyzeStatus;
 import com.starrocks.statistic.BasicStatsMeta;
 import com.starrocks.statistic.HistogramStatsMeta;
+import com.starrocks.statistic.NativeAnalyzeStatus;
 import com.starrocks.system.Backend;
 import com.starrocks.system.ComputeNode;
 import com.starrocks.system.Frontend;
@@ -839,12 +840,12 @@ public class EditLog {
                     break;
                 }
                 case OperationType.OP_ADD_ANALYZE_STATUS: {
-                    AnalyzeStatus analyzeStatus = (AnalyzeStatus) journal.getData();
+                    NativeAnalyzeStatus analyzeStatus = (NativeAnalyzeStatus) journal.getData();
                     globalStateMgr.getAnalyzeManager().replayAddAnalyzeStatus(analyzeStatus);
                     break;
                 }
                 case OperationType.OP_REMOVE_ANALYZE_STATUS: {
-                    AnalyzeStatus analyzeStatus = (AnalyzeStatus) journal.getData();
+                    NativeAnalyzeStatus analyzeStatus = (NativeAnalyzeStatus) journal.getData();
                     globalStateMgr.getAnalyzeManager().replayRemoveAnalyzeStatus(analyzeStatus);
                     break;
                 }
@@ -1586,11 +1587,15 @@ public class EditLog {
     }
 
     public void logAddAnalyzeStatus(AnalyzeStatus status) {
-        logEdit(OperationType.OP_ADD_ANALYZE_STATUS, status);
+        if (status instanceof NativeAnalyzeStatus) {
+            logEdit(OperationType.OP_ADD_ANALYZE_STATUS, (NativeAnalyzeStatus) status);
+        }
     }
 
     public void logRemoveAnalyzeStatus(AnalyzeStatus status) {
-        logEdit(OperationType.OP_REMOVE_ANALYZE_STATUS, status);
+        if (status instanceof NativeAnalyzeStatus) {
+            logEdit(OperationType.OP_REMOVE_ANALYZE_STATUS, (NativeAnalyzeStatus) status);
+        }
     }
 
     public void logAddBasicStatsMeta(BasicStatsMeta meta) {
