+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/AnalyzeStmtAnalyzer.java
@@ -15,6 +15,7 @@
 
 package com.starrocks.sql.analyzer;
 
+import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
@@ -25,6 +26,7 @@ import com.starrocks.catalog.OlapTable;
 import com.starrocks.catalog.Table;
 import com.starrocks.common.Config;
 import com.starrocks.qe.ConnectContext;
+import com.starrocks.server.CatalogMgr;
 import com.starrocks.sql.ast.AnalyzeHistogramDesc;
 import com.starrocks.sql.ast.AnalyzeStmt;
 import com.starrocks.sql.ast.AnalyzeTypeDesc;
@@ -105,6 +107,14 @@ public class AnalyzeStmtAnalyzer {
 
             analyzeProperties(statement.getProperties());
             analyzeAnalyzeTypeDesc(session, statement, statement.getAnalyzeTypeDesc());
+
+            if (CatalogMgr.isExternalCatalog(statement.getTableName().getCatalog())) {
+                if (statement.isSample()) {
+                    throw new SemanticException("External table %s don't support SAMPLE analyze",
+                            statement.getTableName().toString());
+                }
+                statement.setExternal(true);
+            }
             return null;
         }
 
@@ -113,8 +123,14 @@ public class AnalyzeStmtAnalyzer {
             if (null != statement.getTableName()) {
                 TableName tbl = statement.getTableName();
 
+                if ((Strings.isNullOrEmpty(tbl.getCatalog()) &&
+                        CatalogMgr.isExternalCatalog(session.getCurrentCatalog())) ||
+                        CatalogMgr.isExternalCatalog(tbl.getCatalog())) {
+                    throw new SemanticException("External Table don't support analyze job");
+                }
+
                 if (null != tbl.getDb() && null == tbl.getTbl()) {
-                    Database db = MetaUtils.getDatabase(session, statement.getTableName());
+                    Database db = MetaUtils.getDatabase(session, tbl);
 
                     if (StatisticUtils.statisticDatabaseBlackListCheck(statement.getTableName().getDb())) {
                         throw new SemanticException("Forbidden collect database: %s", statement.getTableName().getDb());
@@ -179,6 +195,10 @@ public class AnalyzeStmtAnalyzer {
         private void analyzeAnalyzeTypeDesc(ConnectContext session, AnalyzeStmt statement,
                                             AnalyzeTypeDesc analyzeTypeDesc) {
             if (analyzeTypeDesc instanceof AnalyzeHistogramDesc) {
+                if (CatalogMgr.isExternalCatalog(statement.getTableName().getCatalog())) {
+                    throw new SemanticException("External table %s don't support histogram analyze",
+                            statement.getTableName().toString());
+                }
                 List<String> columns = statement.getColumnNames();
                 OlapTable analyzeTable = (OlapTable) MetaUtils.getTable(session, statement.getTableName());
 
@@ -229,6 +249,9 @@ public class AnalyzeStmtAnalyzer {
         @Override
         public Void visitDropStatsStatement(DropStatsStmt statement, ConnectContext session) {
             MetaUtils.normalizationTableName(session, statement.getTableName());
+            if (CatalogMgr.isExternalCatalog(statement.getTableName().getCatalog())) {
+                statement.setExternal(true);
+            }
             return null;
         }
