+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/PrivilegeCheckerV2.java
@@ -183,6 +183,7 @@ import com.starrocks.sql.ast.UpdateStmt;
 import com.starrocks.sql.ast.UseCatalogStmt;
 import com.starrocks.sql.ast.UseDbStmt;
 import com.starrocks.sql.ast.UserIdentity;
+import com.starrocks.sql.common.MetaUtils;
 import com.starrocks.statistic.AnalyzeJob;
 import com.starrocks.statistic.AnalyzeManager;
 import com.starrocks.statistic.AnalyzeStatus;
@@ -379,17 +380,17 @@ public class PrivilegeCheckerV2 {
         }
     }
 
-    private static void checkTblPrivilegeForKillAnalyzeStmt(ConnectContext context, Database db,
-                                                            Table table, long analyzeId) {
+    private static void checkTblPrivilegeForKillAnalyzeStmt(ConnectContext context, String catalogName, String dbName,
+                                                            String tableName, long analyzeId) {
+        Database db = MetaUtils.getDatabase(catalogName, dbName);
+        Table table = MetaUtils.getTable(catalogName, dbName, tableName);
         if (db != null && table != null) {
-            if (!PrivilegeActions.checkTableAction(context, db.getOriginName(),
-                    table.getName(), PrivilegeType.SELECT) ||
-                    !PrivilegeActions.checkTableAction(context, db.getOriginName(),
-                            table.getName(), PrivilegeType.INSERT)
+            if (!PrivilegeActions.checkTableAction(context, catalogName, dbName, tableName, PrivilegeType.SELECT) ||
+                    !PrivilegeActions.checkTableAction(context, catalogName, dbName, tableName, PrivilegeType.INSERT)
             ) {
                 throw new SemanticException(String.format(
-                        "You need SELECT and INSERT action on %s.%s to kill analyze job %d",
-                        db.getOriginName(), table.getName(), analyzeId));
+                        "You need SELECT and INSERT action on %s.%s.%s to kill analyze job %d",
+                        catalogName, dbName, tableName, analyzeId));
             }
         }
     }
@@ -399,23 +400,20 @@ public class PrivilegeCheckerV2 {
         AnalyzeStatus analyzeStatus = analyzeManager.getAnalyzeStatus(analyzeId);
         AnalyzeJob analyzeJob = analyzeManager.getAnalyzeJob(analyzeId);
         if (analyzeStatus != null) {
-            long dbId = analyzeStatus.getDbId();
-            long tableId = analyzeStatus.getTableId();
-            Database db = GlobalStateMgr.getCurrentState().getDb(dbId);
-            // If the db or table doesn't exist anymore, we won't check privilege on it
-            if (db != null) {
-                Table table = db.getTable(tableId);
-                checkTblPrivilegeForKillAnalyzeStmt(context, db, table, analyzeId);
+            try {
+                String catalogName = analyzeStatus.getCatalogName();
+                String dbName = analyzeStatus.getDbName();
+                String tableName = analyzeStatus.getTableName();
+                checkTblPrivilegeForKillAnalyzeStmt(context, catalogName, dbName, tableName, analyzeId);
+            } catch (MetaNotFoundException ignore) {
+                // If the db or table doesn't exist anymore, we won't check privilege on it
             }
         } else if (analyzeJob != null) {
             Set<TableName> tableNames = PrivilegeCheckerV2.getAllTableNamesForAnalyzeJobStmt(analyzeJob.getDbId(),
                     analyzeJob.getTableId());
             tableNames.forEach(tableName -> {
-                Database db = GlobalStateMgr.getCurrentState().getDb(tableName.getDb());
-                if (db != null) {
-                    Table table = db.getTable(tableName.getTbl());
-                    checkTblPrivilegeForKillAnalyzeStmt(context, db, table, analyzeId);
-                }
+                checkTblPrivilegeForKillAnalyzeStmt(context, tableName.getCatalog(), tableName.getDb(),
+                        tableName.getTbl(), analyzeId);
             });
         }
     }
