+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/DropStatsStmt.java
@@ -21,7 +21,7 @@ import com.starrocks.sql.parser.NodePosition;
 
 public class DropStatsStmt extends StatementBase {
     private final TableName tbl;
-
+    private boolean isExternal = false;
     public DropStatsStmt(TableName tbl) {
         this(tbl, NodePosition.ZERO);
     }
@@ -33,6 +33,14 @@ public class DropStatsStmt extends StatementBase {
 
     public TableName getTableName() {
         return tbl;
+    }
+
+    public boolean isExternal() {
+        return isExternal;
+    }
+
+    public void setExternal(boolean isExternal) {
+        this.isExternal = isExternal;
     }
 
     @Override
