+++ fe/fe-core/src/main/java/com/starrocks/sql/common/MetaUtils.java
@@ -19,6 +19,7 @@ import com.google.common.collect.Maps;
 import com.starrocks.analysis.Expr;
 import com.starrocks.analysis.TableName;
 import com.starrocks.catalog.Database;
+import com.starrocks.catalog.InternalCatalog;
 import com.starrocks.catalog.Table;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.ErrorCode;
@@ -78,9 +79,12 @@ public class MetaUtils {
     }
 
     public static Database getDatabase(ConnectContext session, TableName tableName) {
-        Database db = session.getGlobalStateMgr().getDb(tableName.getDb());
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(session.getCurrentCatalog());
+        }
+        Database db = session.getGlobalStateMgr().getMetadataMgr().getDb(tableName.getCatalog(), tableName.getDb());
         if (db == null) {
-            throw new SemanticException("Database %s is not found", tableName.getDb());
+            throw new SemanticException("Database %s is not found", tableName.getCatalogAndDb());
         }
         return db;
     }
@@ -94,25 +98,25 @@ public class MetaUtils {
     }
 
     public static Table getTable(TableName tableName) {
-        Database db = GlobalStateMgr.getCurrentState().getDb(tableName.getDb());
-        if (db == null) {
-            throw new SemanticException("Database %s is not found", tableName.getDb());
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(InternalCatalog.DEFAULT_INTERNAL_CATALOG_NAME);
         }
-        Table table = db.getTable(tableName.getTbl());
+        Table table = GlobalStateMgr.getCurrentState().getMetadataMgr().getTable(tableName.getCatalog(),
+                tableName.getDb(), tableName.getTbl());
         if (table == null) {
-            throw new SemanticException("Table %s is not found", tableName.getTbl());
+            throw new SemanticException("Table %s is not found", tableName);
         }
         return table;
     }
 
     public static Table getTable(ConnectContext session, TableName tableName) {
-        Database db = session.getGlobalStateMgr().getDb(tableName.getDb());
-        if (db == null) {
-            throw new SemanticException("Database %s is not found", tableName.getDb());
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(session.getCurrentCatalog());
         }
-        Table table = db.getTable(tableName.getTbl());
+        Table table = session.getGlobalStateMgr().getMetadataMgr().getTable(tableName.getCatalog(),
+                tableName.getDb(), tableName.getTbl());
         if (table == null) {
-            throw new SemanticException("Table %s is not found", tableName.getTbl());
+            throw new SemanticException("Table %s is not found", tableName.toString());
         }
         return table;
     }
